# api/openapi.yaml
openapi: 3.1.0
info:
  title: skali API
  description: |
    Client-facing API of skali. Authentication is email+password with
    optional TOTP two-factor; sessions are opaque bearer tokens. The same
    API serves every client: the web BFF, the `skali` CLI, and future
    native apps.

    There is no registration endpoint: users are created by an instance
    admin (`/v1/users`) or by the server operator (`skalid user create`).
    Every user carries an instance-wide role, `admin` (everything,
    including users, nodes, and instance settings) or `member` (nothing
    until granted). Product access is per project and per environment on
    one ladder of roles, `none < read < deploy < maintain < admin`: a
    project membership is the member's default role on every environment
    of the project, a per-environment cell overrides it for one user, and
    an environment ceiling (`max_role`) caps what members inherit.
    Non-members do not see a project at all (`404`); a locked environment
    (`none`) is listed by name but its contents answer `403 forbidden`
    naming the required role. Project and environment payloads carry the
    caller's `access` so clients can disable and explain without guessing.
    All error responses share the `Error` envelope; clients should branch
    on `error.code`, not on messages.

    Security-critical operations sit behind sudo mode: they require the
    session to have proven the user's identity recently (within the
...

...

...
  version: 0.1.0
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/projects/{id}/environments:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: createEnvironment
      summary: Create an environment
      description: |
        Project maintain and up. `priority: high` needs an instance admin.
        A high environment starts with `max_role: read`; a normal one with
        `max_role: admin`. The creator gets an explicit admin cell when
        they are below project admin.
      tags: [environments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  pattern: "^[a-z][a-z0-9-]{0,62}$"
                priority:
                  type: string
                  enum: [normal, high]
                  default: normal
      responses:
        "201":
          description: Environment created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvironmentResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    get:
      operationId: listEnvironments
      summary: List environments of a project
      tags: [environments]
      responses:
        "200":
          description: |
            The project's environments, sorted by name; locked ones carry
            id, name, and `access: none` only.
          content:
            application/json:
              schema:
                type: object
                required: [environments]
                properties:
                  environments:
                    type: array
                    items: { $ref: "#/components/schemas/Environment" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/environments/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getEnvironment
      summary: Get an environment
      description: |
        Project read. A locked environment answers the minimal shape
        (`access: none`).
      tags: [environments]
      responses:
        "200":
          description: The environment.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvironmentResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateEnvironmentSettings
      summary: Update environment settings
      description: |
        Environment admin, sudo mode. Partial: absent properties keep
        their value. Raising `priority` to `high` needs an instance admin;
        `promote_from` must name other environments of the project.
      tags: [environments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: At least one property must be present.
              properties:
                max_role: { $ref: "#/components/schemas/AccessRole" }
                deploy_policy:
                  type: string
                  enum: [direct, promote-only]
                promote_from:
                  type: array
                  items: { type: string }
                priority:
                  type: string
                  enum: [normal, high]
      responses:
        "200":
          description: The environment with its updated settings.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnvironmentResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteEnvironment
      summary: Delete an environment
      description: |
        Destructive: removes the environment with its values, revisions,
        and history. Requires recent authentication (sudo mode).
      tags: [environments]
      responses:
        "204":
          description: Environment deleted.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/ReauthRequired" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque session token from login / 2FA verify.
  responses:
    BadRequest:
      description: Malformed request body or failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing/invalid token, wrong credentials, or wrong code.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: |
        The caller is authenticated but lacks the role the operation needs
        (`error.code = forbidden`; the message names it, for example
        "maintain on environment production required"), or, on
        sudo-mode-gated operations, recent authentication is required
        (`error.code = reauth_required`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: |
        The resource does not exist, or the caller is not a member of the
        project it belongs to (the two are indistinguishable by design).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: The resource is in a state that does not allow the operation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ReauthRequired:
      description: |
        Recent authentication required (`error.code = reauth_required`):
        the session's sudo-mode window has lapsed. Confirm identity at
        `/v1/auth/reauth` and retry the original request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    AccessRole:
      type: string
      description: |
        One step of the access ladder, each including everything below it.
        `none` locks an environment (listed by name, contents refused);
        `read` sees everything; `deploy` changes only what code runs
        (promote, rollback, restart, cancel, backup create, direct deploys
        whose definition is unchanged); `maintain` owns the blast radius of
        skali.yml (definition changes, values, restore, exec, credential
        reveal); `admin` manages settings, other users' roles, delete and
        teardown. Project memberships use `read` and up.
      enum: [none, read, deploy, maintain, admin]
...
    EnvironmentSettings:
      type: object
      description: |
        Server-side environment settings. `max_role` caps the role members
        inherit from their project role (cells are not capped).
        `deploy_policy` and `promote_from` are the protection policy:
        `promote-only` means the running revision only changes through
        promotion from an allowed source, rollback, or an explicit bypass.
        `priority` marks environments that keep running when resources are
        tight; creating with or raising to `high` is instance-admin only.
      required: [max_role, deploy_policy, promote_from, priority]
      properties:
        max_role: { $ref: "#/components/schemas/AccessRole" }
        deploy_policy:
          type: string
          enum: [direct, promote-only]
        promote_from:
          type: array
          description: Environment names allowed as promotion sources; empty means any environment of the project.
          items: { type: string }
        priority:
          type: string
          enum: [normal, high]
...
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - bad_request
            - invalid_credentials
            - invalid_token
            - invalid_code
            - forbidden
            - reauth_required
            - not_found
            - conflict
            - version_conflict
            - invalid_manifest
            - rate_limited
            - node_unreachable
            - no_ready_pod
            - registry_disabled
            - registry_unavailable
            - internal
        message:
          type: string
    Error:
      type: object
      required: [error]
      properties:
        error: { $ref: "#/components/schemas/ErrorDetail" }
...
    Environment:
      type: object
      description: |
        An environment as the caller sees it. `access` is the caller's
        effective role; a locked environment (`none`) carries id, project,
        name, and access only.
      required: [id, project_id, name, access]
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        name:
          type: string
        access: { $ref: "#/components/schemas/AccessRole" }
        created_at:
          type: string
          format: date-time
        settings: { $ref: "#/components/schemas/EnvironmentSettings" }
...
    EnvironmentResponse:
      type: object
      required: [environment]
      properties:
        environment: { $ref: "#/components/schemas/Environment" }
...